requirements.txt
pyyaml
openai
PyPDF2
